+++ src/main/java/com/cristianlima/webapp/biblioteca/service/LibroService.java
@@ -7,12 +7,19 @@ import org.springframework.stereotype.Service;
 
 import com.cristianlima.webapp.biblioteca.model.Libro;
 import com.cristianlima.webapp.biblioteca.repository.LibroRepository;
+import com.cristianlima.webapp.biblioteca.util.LibreriaAlert;
+import com.cristianlima.webapp.biblioteca.util.MethodType;
+
+import javafx.scene.control.ButtonType;
 
 @Service
-public class LibroService implements ILibroService{
+public class LibroService implements ILibroService {
 
     @Autowired
     LibroRepository libroRepository;
+
+    @Autowired
+    LibreriaAlert libreriaAlert;
 
     @Override
     public List<Libro> listarLibros() {
@@ -20,8 +27,23 @@ public class LibroService implements ILibroService{
     }
 
     @Override
-    public Libro guardarLibro(Libro libro) {
-        return libroRepository.save(libro);
+    public Libro guardarLibro(Libro libro, MethodType methodType) {
+        try {
+            if (methodType == MethodType.POST) {
+                libreriaAlert.mostrarAlertaInfo(401);
+                return libroRepository.save(libro);
+            } else {
+                if (libreriaAlert.mostrarAlertaConfirmacion(106).get() == ButtonType.OK) {
+                    libreriaAlert.mostrarAlertaInfo(401);
+                    return libroRepository.save(libro);
+                }
+            }
+        } catch (Exception e) {
+            libreriaAlert.mostrarAlertaInfo(404);
+        }
+        return libro;
+        
+
     }
 
     @Override
@@ -31,13 +53,19 @@ public class LibroService implements ILibroService{
 
     @Override
     public void eliminarLibro(Libro libro) {
-        libroRepository.delete(libro);
+        try {
+            if (libreriaAlert.mostrarAlertaConfirmacion(405).get() == ButtonType.OK) {
+                libroRepository.delete(libro);
+            }
+        } catch (Exception e) {
+            libreriaAlert.mostrarAlertaInfo(404);
+        }
     }
 
     @Override
     public void actualizarDisponibilidad(Libro libro, Boolean disponibilidad) {
-            libro.setDisponibilidad(disponibilidad);
-            guardarLibro(libro);
+        libro.setDisponibilidad(disponibilidad);
+        libroRepository.save(libro);
     }
-    
+
 }
